openai extension: wrong frequency_penalty type

+++ extensions/openai/typing.py
@@ -55,7 +55,7 @@ class CompletionRequest(GenerationOptions):
     logprobs: int | None = None
     max_tokens: int | None = 16
     n: int | None = 1
-    presence_penalty: int | None = 0
+    presence_penalty: float | None = 0
     stop: str | List[str] | None = None
     stream: bool | None = False
     suffix: str | None = None
@@ -82,7 +82,7 @@ class ChatCompletionRequest(GenerationOptions):
     logit_bias: dict | None = None
     max_tokens: int | None = None
     n: int | None = 1
-    presence_penalty: int | None = 0
+    presence_penalty: float | None = 0
     stop: str | List[str] | None = None
     stream: bool | None = False
     temperature: float | None = 1
